requirements.txt
streamlit
tensorflow
numpy
pillow
langchain
langchain-classic
langchain-google-genai
langchain-chroma
chromadb
gdown